Add Notion API doc, update core doc and path modifications
Added a new Notion API documentation file and included it in the main index. Updates made to the core documentation file to include new docs. Paths in the conf.py file updated to include the 'wolf' directory. New imports added to the wolf __init__.py file. Format of the .gitmodules file was updated for better readability. These changes aim to improve documentation and enhance code understandability.