Add test flag for Notion initialisation
Added an optional parameter 'test' set to False by default, to the Notion class during initialisation. This flag is used to prevent actual API calls when running unit tests. If the flag is True, 'test' is used as the token, else the token is read from the "token.json" file. This adjustment ensures that real API calls are not triggered during testing.