Add requirements.txt file and update README files
This commit includes the creation of a requirements.txt file used to build the docs on Read the Docs, advising the use of pyproject.toml instead. It also tidies up the layout in README.md, making it more readable by removing unnecessary line breaks. Furthermore, it updates core/README.md to include the badge for 'Publish to Test PyPI'.