Remove sync_calendar module and add ExampleJob
The sync_calendar module was deleted as it's no longer required due to some changes in the application's design. Instead, a new job called ExampleJob was introduced under example_job.py, the frequency of which is set to run every 5 seconds. The job will randomly return a SUCCESS or ERROR state to simulate different situations for testing purposes. The debug mode was also set to false in the main.py.