Fix incorrect pip command and add package extraction
The former deployment script had an error in pip install command, it was saying 'pip install pip' instead of just using 'pip install'. A section was also added to extract the tar.gz distribution package to a temporary directory, to move contents to the final directory, and to remove the temporary directory. Deactivation of python virtual environment after installing the dependencies was also added for cleanup.