Remove old deployment workflow and update repository owner
Deleted the outdated deployment workflow from GitHub Actions as it's no longer applicable. Transferred the repository ownership from EirLab to Sébastien Delpeuch which is reflected in the changed copyright information and repository URL in documentation and README. The URL of the 'core' submodule also has been updated to match the new owner's repository. These changes simplify the repository management and reflect the actual project's state and ownership.