Update copyright in conf.py and project description in README
Changed copyright attribution from 'EirLab Community' to 'Sébastien Delpeuch' in the conf.py file. The description in the README file has been revised to provide a more general overview of the project, and instructions for cloning the repository, installation, and configuration have been updated accordingly. This alteration reflects the shift in project ownership and direction.