Adjust update frequency and enhance JSON validation
This commit changes the update frequency of the Notion-to-LaTeX conversion from every 1 minute to every 5 minutes to lighten the load on the server. Moreover, an additional validation step is added in the process of converting Notion to LaTeX. Now, if "email" or "name" present in the parameter dictionary, the function will return None. This step prevents potential crashes when these fields are included. Lastly, the log has been changed to be directed to /dev/null to avoid unnecessary accumulation of log files.