Add import difflib, implement check_diff, modify publish_compiled in notion_latex.py
Added import for difflib to allow for file comparisons. Implemented a check_diff method to compare two files and return True if there is a difference, used to identify changes before publishing. Refactored publish_compiled to use check_diff to avoid redundant compilations and publications of unchanged LaTeX. This change improves efficiency and performance of the program.