Correct automatic token injection
Introduced a token injection directly in the deployment workflow, after deactivating the virtual environment. The purpose of the injected tokens is to facilitate the installation process. The tokens were added to the 'install.py' python command to streamline the cloud-based publishing systems. This commit is aimed to enhance the security and efficiency of application deployment under the 'deploy_sonu_dev_gazebo.yml' GitHub action.