Update tags in GitHub workflow file
This commit updates the tag value from "ci" to "wolf" in the .github/workflows/deploy_sonu_dev_gazebo.yml file. The change was necessary to align with the amalgamated tracking system for workflow processes. Just to note, the 'tags' attribute helps in classifying or differentiating specific workflows for better organization and management.