Move Tailscale action to the deploy job in GitHub workflow
In this commit, we removed the standalone 'tailscale' job in the Github workflow, specifically in the .github/workflows/deploy_sonu_dev_gazebo.yml file and incorporated it into the 'deploy' job. It was done to streamline the workflow by reducing unnecessary segmentation of jobs. This change enhances the overall efficiency without affecting the workflow process as Tailscale continues to function as expected.