Update GitHub workflow to optimize LaTeX operations
This commit makes changes in the `notion_latex.yml` workflow. The TeX Live setup action is now integrated directly into the workflow that runs when a push is made. This simplifies and enhances the TeX Live installation process. We also removed some unnecessary code related to PDF generation, which is no longer needed thanks to the new updates. We now publish the output as an artifact directly to Github. The modifications aim to increase the efficiency and streamline the overall continuous integration process.