Refactor notion_latex.py for simplicity and stability
The difflib and Github libraries were removed due to unused imports, refactor run_command to be a static method instead of instance method. Removed method check_diff because it's no longer needed for our functionality. We no longer check for difference between the current and last file for compilation. We also removed the 'link' parameter from update_notion method as it was unused. More changes were made for cleaning and simplifying the code, enhancing overall readability and maintainability.